check number of slices correctly before print v3bw
aggregator should check the number of slices actually being
measured according to the number of measurable relays as found
in the descriptor list from tor, instead of the total number of
relays found in the initial input v3bw file. the number of
measurable relays could be lower because they might not
have the FAST flag, or may not have all bootstrapped correctly.