Fix building without tor tracing
Before, *.c files in the trace directory would always be built,
which was a problem since they depended on options being defined
in other files that aren't always defined.

This adds the ability to conditionally add dependencies to
shadow-plugin-tor. Just add them to the additional_sources variable
when your condition is defined. See TOR_TRACING example.