Update Tor to latest stable version
The default version of Tor has been updated to the latest stable
release 0.3.2.9.

Also fixed an issue in Tor commit
72b5e4a2db4282002fe50e11c2b8a79e108d30f8
that was added between 0.3.3.0 and the as-yet unreleased 0.3.3.1
where the git revision string definition was moved and is no
longer necessary to get defined in shadow. This fix should
automatically work as of 0.3.3.1-alpha, whenever that gets
released.

fixes #44